+++ CHANGELOG.rst
@@ -3,6 +3,20 @@
 Changelog
 #########
 
+0.4 November 7, 2016
+====================
+
+* Properly continue the allauth login workflow after successful 2FA login, e.g.
+  send allauth signals
+* Support using ``MIDDLEWARE`` setting with Django 1.10.
+* Support customer ``USERNAME_FIELD`` on the auth model.
+
+0.3.2 October 26, 2016
+======================
+
+* Fix an error when hitting the TwoFactorBackupTokens view as a non-anonymous
+  user.
+
 0.3.1 October 5, 2016
 =====================